refactor(ra): resolve a game's RetroAchievements id in one place
RetroAchievementsHelper._findGameId and RetroAchievementsResolver
.resolveGameId were the same four-strategy pipeline written twice, with
separately-written normalization, and they had already drifted: the
resolver rejected a game id of 0 at every strategy while the helper
accepted it, and the resolver's filename and title normalization was
extracted and tested while the helper's was inline and not. The manual
match fix in #361 had to be written twice for the same reason, and the
plan calls this out as a prerequisite for disc hashing — a change landing
in one copy leaves the secondary display and the main UI disagreeing
about which game is running.

RetroAchievementsMatcher now holds the whole pipeline: hash resolution,
the manual pick, the hash lookup, the filename fallback and the history
match. The helper keeps what is genuinely the main display's own — the
"All" mode folder choice, the provider call and badge-image eviction —
and the resolver keeps the snapshot the secondary panel renders.

Behaviour is unchanged except that the helper now rejects a game id of 0
at every strategy, which the resolver already did. The moved
sanitizeRomName/normalizeTitle tests move with them.

AI-Assisted: Claude:Opus-5
