+++ lib/services/retro_achievements_resolver.dart
@@ -1,13 +1,8 @@
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
-
 import '../models/game_model.dart';
-import '../models/retro_achievements_summary.dart';
 import '../models/secondary_achievement_item.dart';
 import '../providers/retro_achievements_provider.dart';
-import '../repositories/retro_achievements_repository.dart';
 import 'logger_service.dart';
+import 'retro_achievements_matcher.dart';
 import 'retroachievements_hash_service.dart';
 
 /// A condensed snapshot of a game's RetroAchievements progress, ready to be
@@ -53,107 +48,14 @@ class SecondaryAchievementsSnapshot {
       achievements.where((a) => a.earned).map((a) => a.id).toSet();
 }
 
-/// Resolves a local [GameModel] to its RetroAchievements game id and fetches a
-/// condensed progress snapshot for the secondary display.
+/// Fetches a condensed RetroAchievements progress snapshot for the secondary
+/// display.
 ///
-/// The resolution strategy mirrors the one used by the game-details
-/// achievements view (`game_details_card_list.dart`): exact MD5-hash match
-/// against the local RA database, then a sanitized-filename lookup, then a
-/// heuristic match against the user's recently-played history. The underlying
-/// hashing and database lookups are reused from [RetroAchievementsHashService]
-/// and [RetroAchievementsRepository].
+/// Identification is [RetroAchievementsMatcher]'s job and is shared with the
+/// main display, so the two screens cannot disagree about which game is
+/// running. This adds only the snapshot the secondary panel renders.
 class RetroAchievementsResolver {
   static final _log = LoggerService.instance;
-
-  /// Maximum ROM size for which a fallback MD5 hash is generated, mirroring the
-  /// limit applied in the game-details achievements flow.
-  static const int _maxHashFileSize = 512 * 1024 * 1024;
-
-  /// Resolves the MD5 hash used for RA matching, generating one if absent.
-  static Future<String?> resolveMd5Hash(
-    GameModel game, {
-    required bool hasSpecificGenerator,
-  }) async {
-    var md5Hash = game.raHash;
-    if (md5Hash != null && md5Hash.isNotEmpty) return md5Hash;
-
-    if (hasSpecificGenerator) {
-      // Core systems (e.g. PSX, GBA): always generate for precise matching.
-      return RetroAchievementsHashService.generateHashForGame(game);
-    }
-
-    // Fallback systems: only hash reasonably sized files to keep launch snappy.
-    final romPath = game.romPath;
-    if (romPath == null) return null;
-    final file = File(romPath);
-    if (!await file.exists()) return null;
-    if (await file.length() >= _maxHashFileSize) return null;
-    return RetroAchievementsHashService.generateHashForGame(game);
-  }
-
-  /// Resolves the RA game id for [game] using hash, filename, then history.
-  static Future<int?> resolveGameId({
-    required GameModel game,
-    required String systemFolderName,
-    required RetroAchievementsUserSummary? summary,
-    required bool hasSpecificGenerator,
-    String? md5Hash,
-  }) async {
-    // Strategy 0: a match the user chose by hand wins over every automatic
-    // strategy, here as well as in RetroAchievementsHelper — otherwise the
-    // secondary display and the main UI disagree about which game is running.
-    final romPath = game.romPath;
-    if (romPath != null && romPath.isNotEmpty) {
-      try {
-        final manualId = await RetroAchievementsRepository.getManualRomRaGameId(
-          romPath,
-        );
-        if (manualId != null && manualId != 0) return manualId;
-      } catch (e) {
-        _log.e('RA resolver: manual match lookup failed: $e');
-      }
-    }
-
-    // Strategy 1: exact hash match against the local RA database.
-    if (md5Hash != null && md5Hash.isNotEmpty) {
-      try {
-        final gameId = await RetroAchievementsRepository.findGameIdByHash(
-          md5Hash,
-        );
-        if (gameId != null && gameId != 0) return gameId;
-      } catch (e) {
-        _log.e('RA resolver: hash lookup failed: $e');
-      }
-    }
-
-    // Hash-specific systems must match by hash only to avoid false positives.
-    if (hasSpecificGenerator) return null;
-
-    // Strategy 2: sanitized filename match.
-    try {
-      final gameId = await RetroAchievementsRepository.findGameIdByFilename(
-        systemFolderName,
-        sanitizeRomName(game.romname),
-      );
-      if (gameId != null && gameId != 0) return gameId;
-    } catch (e) {
-      _log.e('RA resolver: filename lookup failed: $e');
-    }
-
-    // Strategy 3: heuristic match against recently-played history.
-    try {
-      final normalizedLocal = normalizeTitle(game.name);
-      for (final recent in summary?.recentlyPlayed ?? const []) {
-        if (normalizeTitle(recent.title) == normalizedLocal) {
-          return recent.gameId;
-        }
-      }
-    } catch (e) {
-      _log.e('RA resolver: history lookup failed: $e');
-    }
-
-    return null;
-  }
 
   /// Resolves and fetches a condensed achievements snapshot for [game].
   ///
@@ -172,11 +74,11 @@ class RetroAchievementsResolver {
           RetroAchievementsHashService.hasSpecificHashGenerator(
             game.systemFolderName,
           );
-      final md5Hash = await resolveMd5Hash(
+      final md5Hash = await RetroAchievementsMatcher.resolveMd5Hash(
         game,
         hasSpecificGenerator: hasSpecificGenerator,
       );
-      final gameId = await resolveGameId(
+      final gameId = await RetroAchievementsMatcher.resolveGameId(
         game: game,
         systemFolderName: systemFolderName,
         summary: provider.userSummary,
@@ -230,29 +132,5 @@ class RetroAchievementsResolver {
       _log.e('RA resolver: failed to fetch snapshot for ${game.name}: $e');
       return null;
     }
-  }
-
-  /// Strips the file extension and any `(...)` / `[...]` tags (region, revision,
-  /// dump flags) from a ROM filename, leaving a bare title for RA lookup.
-  @visibleForTesting
-  static String sanitizeRomName(String romname) {
-    final withoutExtension = romname.contains('.')
-        ? romname.substring(0, romname.lastIndexOf('.'))
-        : romname;
-    return withoutExtension
-        .replaceAll(RegExp(r'\([^)]*\)'), '')
-        .replaceAll(RegExp(r'\[[^\]]*\]'), '')
-        .trim();
-  }
-
-  /// Normalizes a title for loose equality: lowercased, punctuation removed,
-  /// whitespace collapsed. Used to match local names against RA history.
-  @visibleForTesting
-  static String normalizeTitle(String value) {
-    return value
-        .toLowerCase()
-        .replaceAll(RegExp(r'[^\w\s]'), '')
-        .replaceAll(RegExp(r'\s+'), ' ')
-        .trim();
   }
 }
